add migration previews
