+++ Sources/FluentKit/Database/FluentDatabaseID.swift
@@ -1,0 +1,6 @@
+public struct FluentDatabaseID: Hashable, Codable {
+    public let string: String
+    public init(string: String) {
+        self.string = string
+    }
+}
